backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
openai==0.28.0
python-dotenv==1.0.1
gunicorn==21.2.0
pyjwt==2.3.0
uuid==1.30
Flask-SQLAlchemy==3.1.0
psycopg2-binary